+++ ext_test_missingfn_1062/source.lua
@@ -1,0 +1,1 @@
+return { popular_manga = function() end }